tests/wpt/tests/html/semantics/popovers/popover-types-with-hints-and-anchor.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popover.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
  function getPopovers() {
    return Array.from(document.currentScript.parentElement.querySelectorAll('[popover]'));
  }
  function assertState(expectedState,description) {
    description = description || 'Error';
    const popovers = getPopovers();
    const n = popovers.length;
    assert_equals(expectedState.length,n,'Invalid expected state length');
    for(let i=0;i<n;++i) {
      const html = '<' + popovers[i].outerHTML.split('<')[1] + '</div>';
      assert_equals(popovers[i].matches(':popover-open'),expectedState[i],`${description}, index ${i} (${html})`);
    }
  }
</script>

<div>
  <div popover>popover 1
    <div popover>popover 2
      <p id=anchorid>Anchor</p>
      <div popover>popover 3</div>
    </div>
  </div>
  <div popover=hint anchor=anchorid>Hint anchored to popover</div>
  <script>
  {
    const popover1 = getPopovers()[0];
    const popover2 = getPopovers()[1];
    const popover3 = getPopovers()[2];
    const hint = getPopovers()[3];
    test(() => {
      assertState([false,false,false,false]);
      popover1.showPopover();
      popover2.showPopover();
      popover3.showPopover();
      assertState([true,true,true,false]);
      hint.showPopover(); // Because hint is nested in popover2, popover3 should be hidden
      assertState([true,true,false,true]);
      popover1.hidePopover(); // Should close the hint, which is anchored to popover2
      assertState([false,false,false,false]);
    },'hint is not closed by pre-existing auto');
  }
  </script>
</div>
