third_party/blink/web_tests/virtual/html-anchor-attribute-disabled/external/wpt/html/semantics/popovers/popover-types-with-hints-and-anchor.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] hint is not closed by pre-existing auto
  assert_equals: Error, index 2 (<div popover="">popover 3</div>) expected false but got true
Harness: the test ran to completion.
